Add handling to save new OrderState values when a Execution arrives via websocket from the exchange.
This runs in the background to avoid some Rust thread-safety
messiness, which then necessitated putting a Mutex around the database
updates.  This avoid excess optimistic locking exceptions.  Some of
those are caused by a current defect in OpenExchange that sends
duplicate "Filled" updates.  This happen so quickly that they are
essentially guaranteed to be handled simultaneously in OpenBroker.