Add management of positions and balances when executions are received, including saving into the database.  All closing gains are computed using average opening price, not by lot matching.
Build and use a wrapper, AccountUpdate, for all account-level updates,
allowing the receiver to know which type of message has arrived.  When
updates are sent based on executions, multiple messages are combined,
for more efficient transmission and handling.

Add version_number to OrderStates sent via the REST API and via
websockets, allowing the client to discard outdated messages.
version_numbers are also provided in position and balance messages.