Add ability to request initial data snapshots via websockets.
By requesting in this way, the client need not make any REST requests
to get an initial state.  More importantly, a client that wants to
handle real-time updates client need only handle websocket messages
for the balance, position, and order data, instead of also needing to
implement handling of the REST responses.

The REST API is maintained for simple snapshot clients and clients
for which polling is more sensible.