Add the ability to dynamically add exchanges and load instruments from exchanges.
Exchange configuration is no longer stored in the environment.
Instead, multiple exchanges can be stored in the database,
added at runtime via an admin-only REST API.  Once the exchange is
added, another admin REST API can be invoked to download the
instruments from the exchange.  The instruments are saved in the
database and also available immediately.

On startup, the exchange table is read, and all available exchanges
are configured, including downloading instruments.